factored day 04 into solution interface
